Add a skeleton mex file with build system.
The build system works under Linux, after plenty of trial and error, since the
CMake documentation was sparse in this regard. Also, a "fork" of
FindMatlab.cmake was necessary to find the Matlab include files and binaries. It
turns out CMake seems to handle mex-File compilation properly (calling mex
correctly, etc, without explicitly telling it to do so).

The build system was tested with a trivial mex-File (an empty function) which
compiled and linked successfully and apparently works fine in Matlab (i.e., it
does nothing).

NOTE: the build system has *not* been adapted for, let alone tested on, Windows
and Mac (yet).