CMake: finalize build system (hopefully).
Corrected some misconceptions in build system (mex is not called, how silly of
me to think so). It turns out "FindMatlab" tried to link to the mex binary
instead of to libmex, resulting in those file format errors:
    /opt/bin/mex: file not recognized: File format not recognized
    ^^^
    (that made me think something was calling mex, error messages are lovely)

Also added mexversion.{c,rc} to the source files, as per the Matlab documentation.

Everything should be setup so that I can finally start with the actual
programming and have everything work under Windows, Mac and Linux with whatever
IDEs/make implementations/etc. CMake supports.