Split libsndfile wrapper off of Dirac wrapper.
I split the libsndfile wrapper off of the Dirac wrapper. This simplifies
developing and testing both and allows me to do more in Matlab (e.g.,
resampling).  Also, TimeStretchDirac was renamed to mDirac in the CMakeLists.txt
file; dirac.{c,def} was renamed to mDirac.{c,def} and now has a matching m-file.