mDirac: correct the array index handling.
I had misunderstood an example in the Matlab documentation and programmed the
interleaved array index handling incorrectly, mixing up the input and output
arrays.  This change apparently makes mDirac actually work correctly for MxN
arrays, assuming M is the number of channels and N the number of samples per
channel.