Expand compile_timestretch.m to work on Linux and Mac.
As an easy to use alternative to the full-fledged SCons build system, there is
now an m-File that compiles via the mex command. While it is much easier to use
and doesn't require additional software, it doesn't offer full control over the
system toolchain, which is the goal of the SCons build system.