TimeStretchDirac: fix sampling rate check.
The check for whether the sampling rate is compatible with Dirac was done before
the audio was read in, which led to an unnecessary warning when passing a file
name without a sampling rate (which is what you would normally do). The check is
now conducted after the file is read, which sets the sampling rate and makes the
isfloat() check unnecessary, which has been removed accordingly.